requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.1
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.1.3
django-colorfield==0.11.0
django-extensions==3.2.3
django-oauth-toolkit==3.0.1
django-random-id-model==0.2.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.15.2
frozenlist==1.5.0
future==1.0.0
geocoder==1.38.1
googlemaps==4.10.0
idna==3.10
jwcrypto==1.5.6
multidict==6.1.0
oauth2_provider==0.0
oauthlib==3.2.2
phonenumbers==8.13.49
pillow==11.0.0
propcache==0.2.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
ratelim==0.1.6
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stripe==11.2.0
twilio==9.3.6
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
